move getSolventList from views to models
no imports from views into models will avoid cyclic import
errors
